refactor: move standard, complex model in one file
