+++ xeofs/models/__init__.py
@@ -1,7 +1,5 @@
-from .eof import EOF
-from .complex_eof import ComplexEOF
-from .mca import MCA
-from .complex_mca import ComplexMCA
-from .rotator import RotatorFactory
-from .eof_rotator import EOFRotator
-from .mca_rotator import MCARotator
+from .eof import EOF, ComplexEOF
+from .mca import MCA, ComplexMCA
+from .rotator_factory import RotatorFactory
+from .eof_rotator import EOFRotator, ComplexEOFRotator
+from .mca_rotator import MCARotator, ComplexMCARotator
